requirements.txt
streamlit
langchain
langchain_community
langchain_google_genai
faiss-cpu
python-dotenv
firecrawl-py